CHROMIUM: seccomp_filter: make inherited filters composable
Right now, inheritance just copies the parent ruleset.  This means
that on inheritance, if a child wants to reduce privilege, it must
walk its inherited set to drop those it doesn't want.

With this change, inherited sets are always left intact and any process
local rules are created in a child seccomp_filter object.  On evaluation,
the child rules are checked first, then all ancestors.  If the system call
is not allowed by any of them, the process will be terminated.

The change makes PR_SECCOMP_GET_FILTER less intuitive since it won't
coalesce all inherited values, but it assures that a child process
can create its ruleset normally, then call prctl(PR_SET_SECCOMP, 13) to
enable it on top of anything that is inherited.

This is similar to namespacing behavior except that the copy-on-update
methodology precludes the parent process from modifying a filter that
has been inherited.

TEST=minijail0 -S support and the seccomp_tests.c in the seccomp filters bugs
BUG=chromium-os:19459

Change-Id: Icd192b75d887fbee09d99fe86cd1a815a8ae673d
Signed-off-by: Will Drewry <wad@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7109
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>