UBUNTU: add a new linux-image-extras package for virtual
Add a new linux-image-extras package for the virtual flavours.
This package contains all of the left over kernel modules which are not
in the main linux-image package.  This allows users to opt-in to a full
server style install in a virtual machine, preventing slow bloat of the
core linux-image install.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>