UBUNTU: debian: add locking to protect debian/files from parallel update
When building with high parallelism we may overlap calls to dh_gencontrol
and/or the .ddeb removal code from the debian/files list.  If these happen
in parallel files may be lost or updates fail leading to build failures or
incomplete packing lists.  Add primative locking to each of these scenarios
using the flock command, utilising a lock file in the debian directory.

Signed-off-by: Andy Whitcroft <apw@canonical.com>