ARM: vexpress: Use FDT data in platform SMP calls
If ct_desc is not set (so a board was booted with DT machine ID)
try to look for "arm,cortex-a9-scu" node and initialize the SCU
using base address in "reg" property. Otherwise assume that
there is no specific SCU initialization required and initialize
cpus based numbers of "cpu" type devices in "cpus" node.

Signed-off-by: Pawel Moll <pawel.moll@arm.com>